Added method to Instrument class to setSellPrice, TDD tests passed.

+++ music_shop_management/Piano.java
@@ -5,7 +5,7 @@ public class Piano extends Instrument implements Playable, Sellable {
 
   int keys;
 
-  public Piano(String make, String material, String colour, String type, int buyPrice, int sellPrice, int keys) {
+  public Piano(String make, String material, String colour, String type, double buyPrice, double sellPrice, int keys) {
     super(make, material, colour, type, buyPrice, sellPrice);
     this.keys = keys;
   }
