+++ music_shop_management/Trumpet.java
@@ -5,7 +5,7 @@ public class Trumpet extends Instrument implements Playable, Sellable {
 
   int valves;
 
-  public Trumpet(String make, String material, String colour, String type, int buyPrice, int sellPrice, int valves) {
+  public Trumpet(String make, String material, String colour, String type, double buyPrice, double sellPrice, int valves) {
     super(make, material, colour, type, buyPrice, sellPrice);
     this.valves = valves;
   }
